Rewrap text at a given width
I have an automation that copies git commits when raising a PR.  The git
commits are wrapped at a given width, but this looks awful in the github
UI.  I wanted a nice way to re-wrap text both inside and outside of my
editor.  Given my love of raycast, I vibe-coded this extension to rewrap
text at any given width.

NOTE this extension currently doesn't respect indentation. If you copy a
JSON blob or a markdown bullet list it'll squash everything to the
side).

NOTE this extension currently respects line breaks (but we probably
don't want this).